Set up macros. Add to readme.
Set up some macros to create functions and comments according to the
style at work. Added some directions in the readme file to remind myself
how to clone the repo. Changed the default width of the gvim window to
be a little more narrow. Changed the default Linux font to monospace.