firefox: force popup menu to use light scheme
Setting `menupop` to use the light color scheme changes the context menu
to match the toolbar popup menu. Without this, navigating to the
bookmarks sidebar and then back to tree tabs changed the popup menu in
the sidebar to dark mode.

Configuring `menupop` alone resulted in some responsive favicons (like
docs.rs) to change to dark theme and serve a light colored favicon.
Setting `:root` to the light scheme fixed this.