nvim, tmux: configure writer environment
Markdown likes to have text with soft line breaks. I tried doing that
with `Wm` by setting the columns of the view port, but it breaks any
time the pane gets resized and isn't a smooth experience when toggling
of write mode. Instead, write mode now soft breaks to the width of the
window. I added a key binding in tmux to create a centered pane that's
126 character wide to get the same effect as the original write mode.

This change is also motivated by the switch to `SilverBullet` for
hosting a personal knowledge system. As a result, also removing the
`obisidan` plugin.